+++ MALClient.Dekstop/UserControls/New/HamburgerControl.xaml.cs
@@ -59,7 +59,7 @@ namespace MALClient.UserControls.New
                     ButtonExpandTopCategoriesOnClick(null, null);
             }
 
-            FeedbackImage.Source = Settings.SelectedTheme == ApplicationTheme.Dark
+            FeedbackImage.Source = Settings.SelectedTheme == (int)ApplicationTheme.Dark
                 ? new BitmapImage(new Uri("ms-appx:///Assets/GitHub-Mark-Light-120px-plus.png"))
                 : new BitmapImage(new Uri("ms-appx:///Assets/GitHub-Mark-120px-plus.png"));
         }
